# airbnb-monolith/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Airbnb API
  version: 1.0.0
  description: |
    REST API for the Airbnb-like platform. Supports listing management, booking lifecycle,
    post-checkout resolutions, and in-app notifications.

    **Authentication:** JWT Bearer tokens. Obtain tokens via `/auth/login`, `/auth/register-user`,
    or `/auth/register-owner`, then pass them as `Authorization: Bearer <token>` on protected endpoints.
servers:
  - url: http://localhost:8085
    description: Local dev server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── SCHEMAS ──────────────────────────────────────────────────────────────
  schemas:
...
    UserRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          description: Alphanumeric, underscores and spaces; 3–50 chars
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Username or email accepted by the @Login validator
        password:
          type: string
          format: password
          minLength: 8
    TokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    ValidateTokenResponse:
      type: object
      properties:
        valid:
          type: boolean
        username:
          type: string
          description: Empty string when token is invalid

    # ── Listing DTOs ──────────────────────────────────────────────────────────
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: object
          nullable: true
        errorType:
          type: string
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          nullable: true
        timestamp:
          type: string
          format: date-time

  # ─── REUSABLE PARAMETERS ──────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Validation or parsing error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /auth/register-user:
    post:
      tags: [Auth]
      summary: Register a new user (guest, ROLE_USER)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created; JWT tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/register-owner:
    post:
      tags: [Auth]
      summary: Register a new owner (ROLE_USER + ROLE_OWNER)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Owner created; JWT tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with username/email and password (JAAS)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated; JWT tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/validate:
    post:
      tags: [Auth]
      summary: Validate a token and extract the subject username
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateTokenResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate the current access token
      responses:
        '200':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ════════════════════════════════════════════════════════════════════════════
  # USERS
  # ════════════════════════════════════════════════════════════════════════════
...